Create 17-array.java
Task

The code in your editor does the following:

Reads an integer from stdin and saves it to a variable,n , denoting some number of integers.
Reads n integers corresponding to a0,....,an-1 from stdin and saves each integer ai to a variable, val.
Attempts to print each element of an array of integers named a .
Write the following code in the unlocked portion of your editor:

Create an array,a , capable of holding n integers.
Modify the code in the loop so that it saves each sequential value to its corresponding location in the array. For example, the first value must be stored in a0, the second value must be stored in a1 , and so on.
Good luck!